0.2.62
Creo un componente para estados vacíos y lo uso en la vista de almacenes.
